Update to Agave v2.2
#### Problem

The Agave and Solana-SDK crates v2.2 are now available, but the nix
derivations are still using v2.1.

This is most interesting because cargo-build-sbf now contains specific
flags to support nix builds!

#### Summary of changes

Bump everything as needed:

* Use v2.2.3 of Agave
* Use v0.31.0 of Anchor
* Use newer nixpkgs and rust-overlay
* Change sdk directory name to platform-tools-sdk
* Remove cargo-build-sbf patch which isn't needed anymore
* Update Anchor CLI patch to use the new `--skip-tools-install` and
  `--no-rustup-override` flags by default
* Update Solana CLI to Rust v1.84.1
* Update Anchor CLI to Rust v1.85.0

Note that the removal of the patch means that build-sbf commands *must*
specify `--skip-tools-install` and `--no-rustup-override`. We could
equally remove the patch for Anchor, in which case people will have to
specify those flags. Let me know what you think!